Namtrinh

+++ rating-review-service/src/main/java/org/example/ratingreviewservice/dto/request/ReviewDTO.java
@@ -4,13 +4,16 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import org.example.ratingreviewservice.entity.ReviewImage;
+
+import java.util.List;
 
 @Data
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
 public class ReviewDTO {
-    private int id;
+    private Long id;
     private int productId;
     private String productName;
     private int rating;
